refactor(mergeScan): moved out of extended directory
BREAKING CHANGE: `mergeScan` has been moved out of `extended` folder and under `operator` folder directly
