--- src/add/operator/extended/mergeScan.ts
@@ -1,7 +1,0 @@
-import {Observable} from '../../../Observable';
-import {mergeScan} from '../../../operator/extended/mergeScan';
-import {KitchenSinkOperators} from '../../../Rx.KitchenSink';
-const observableProto = (<KitchenSinkOperators<any>>Observable.prototype);
-observableProto.mergeScan = mergeScan;
-
-export var _void: void;
