+++ src/operator/mergeScan.ts
@@ -1,11 +1,11 @@
-import {Operator} from '../../Operator';
-import {Observable} from '../../Observable';
-import {Subscriber} from '../../Subscriber';
-import {Subscription} from '../../Subscription';
-import {tryCatch} from '../../util/tryCatch';
-import {errorObject} from '../../util/errorObject';
-import {subscribeToResult} from '../../util/subscribeToResult';
-import {OuterSubscriber} from '../../OuterSubscriber';
+import {Operator} from '../Operator';
+import {Observable} from '../Observable';
+import {Subscriber} from '../Subscriber';
+import {Subscription} from '../Subscription';
+import {tryCatch} from '../util/tryCatch';
+import {errorObject} from '../util/errorObject';
+import {subscribeToResult} from '../util/subscribeToResult';
+import {OuterSubscriber} from '../OuterSubscriber';
 
 export function mergeScan<T, R>(project: (acc: R, x: T) => Observable<R>,
                                 seed: R,
